+++ app/controllers/public/registrations_controller.rb
@@ -28,7 +28,7 @@ class Public::RegistrationsController < Devise::RegistrationsController
   # PUT /resource
   def update
     if update_resource(current_user, user_params)
-      render json: { message: '投稿が成功しました'}, status: :ok
+      render json: { message: 'ユーザー情報が更新されました' }, status: :ok
     else
       render json: { errors: current_user.errors.full_messages }, status: :unprocessable_entity
     end
@@ -43,10 +43,10 @@ class Public::RegistrationsController < Devise::RegistrationsController
   end
 
   def update_resource(resource, params)
-    if params[:password].present? || params[:password_confirmation].present?
+    if params[:password].present? && params[:password_confirmation].present?
       resource.update_with_password(params)
     else
-      resource.update_without_password(params)
+      resource.update_without_password(params.except(:current_password, :password, :password_confirmation))
     end
   end
