+++ app/controllers/public/users_controller.rb
@@ -33,7 +33,7 @@ class Public::UsersController < ApplicationController
     if @user.destroy
       render json: { message: '投稿が成功しました', redirect_url: root_path }, status: :ok
     else
-      render json: { error: '投稿に失敗しました', details: @user.errors.full_messages }, status: :unprocessable_entity
+      render json: { error: @user.errors.full_messages }, status: :unprocessable_entity
     end
   end
